requirements.txt
jsonpath==0.82
requests==2.22.0
PyMySQL==0.9.3
PyYAML==5.3.1
